app/db_management/db.py
import mysql.connector

def get_db_connection():
    conn = mysql.connector.connect(
        host='localhost',
        user='root',
        password='',
        database='aicg',
        port = 3307
    )
    return conn