app/services/hf_test.py
from groq import Groq
from dotenv import load_dotenv
import os

load_dotenv()

client = Groq(api_key=os.getenv("GROQ_API_KEY"))

try:
    response = client.chat.completions.create(
        model="llama-3.3-70b-versatile",
        messages=[
            {"role": "system", "content": "You are an expert curriculum designer."},
            {"role": "user", "content": "Generate a short Python course outline."}
        ],
        temperature=0.7,
    )

    print("✅ Success\n")
    print(response.choices[0].message.content)

except Exception as e:
    print("❌ Error:", e)