requirements.txt
accelerate==1.12.0
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.10.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distro==1.9.0
Django==4.2
filelock==3.20.0
Flask==3.1.2
fsspec==2025.10.0
git-filter-repo==2.47.0
groq==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
MarkupSafe==3.0.3
mpmath==1.3.0
mysql-connector-python==9.6.0
mysqlclient==2.2.7
networkx==3.6.1
numpy==2.4.2
packaging==25.0
pillow==11.3.0
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
regex==2026.2.19
requests==2.32.5
safetensors==0.7.0
sentencepiece==0.2.1
setuptools==82.0.0
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0+cpu
tqdm==4.67.1
transformers==5.2.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.46.3
